backend/variant-embedding-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
torch==2.3.1
torchvision==0.18.1
pillow==10.4.0
requests==2.32.3
